requirements.txt
    kabbes_py_starter
    kabbes_dir_ops   
